# src/routing/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"IErrorResponseDto": {
				"description": "Тело любого ответа с ошибкой (`status >= 400`).",
				"properties": {
					"status": {
						"type": "number",
						"format": "double",
						"description": "HTTP-статус."
					},
					"code": {
						"type": "string",
						"description": "Машинный код: доменный (`USER_EMAIL_TAKEN`) или по статусу (`NOT_FOUND`,\n`VALIDATION_ERROR`, `INTERNAL_ERROR`). Стабилен — на него опирается клиент."
					},
					"message": {
						"type": "string",
						"description": "Человекочитаемое сообщение; для 5xx — общее, без внутренних деталей."
					},
					"details": {
						"description": "Подробности: для VALIDATION_ERROR — поле → сообщение."
					},
					"requestId": {
						"type": "string",
						"description": "Идентификатор запроса — для обращения в поддержку и поиска в логах."
					},
					"stack": {
						"type": "string",
						"description": "Stack trace — только вне production."
					}
				},
				"required": [
					"status",
					"code",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"UUID": {
				"type": "string",
				"description": "Идентификатор сущности в пути или теле запроса. tsoa проверяет формат до\nконтроллера: неверный uuid → 422, а не ошибка Postgres.",
				"pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
			},
...
			"EWgPeerDisabledReason": {
				"description": "Причина, по которой пир выключен.",
				"enum": [
					"manual",
					"expired"
				],
				"type": "string"
			},
			"WgPeerDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"interfaceId": {
						"type": "string"
					},
					"interfaceName": {
						"type": "string",
						"nullable": true
					},
					"nodeId": {
						"type": "string",
						"nullable": true
					},
					"nodeName": {
						"type": "string",
						"nullable": true
					},
					"userId": {
						"type": "string",
						"nullable": true
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"publicKey": {
						"type": "string"
					},
					"hasPrivateKey": {
						"type": "boolean",
						"description": "Можно ли выпустить конфиг/QR (приватный ключ хранится)."
					},
					"hasPresharedKey": {
						"type": "boolean"
					},
					"addressV4": {
						"type": "string"
					},
					"addressV6": {
						"type": "string",
						"nullable": true
					},
					"clientAllowedIPs": {
						"type": "string"
					},
					"clientDns": {
						"type": "string",
						"nullable": true
					},
					"clientMtu": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"persistentKeepalive": {
						"type": "number",
						"format": "double"
					},
					"enabled": {
						"type": "boolean"
					},
					"disabledReason": {
						"allOf": [
							{
								"$ref": "#/components/schemas/EWgPeerDisabledReason"
							}
						],
						"nullable": true
					},
					"expiresAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"lastHandshakeAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"lastEndpoint": {
						"type": "string",
						"nullable": true
					},
					"isOnline": {
						"type": "boolean",
						"description": "Handshake свежее 3 минут."
					},
					"rxBytesTotal": {
						"type": "number",
						"format": "double"
					},
					"txBytesTotal": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"interfaceId",
					"interfaceName",
					"nodeId",
					"nodeName",
					"userId",
					"name",
					"description",
					"publicKey",
					"hasPrivateKey",
					"hasPresharedKey",
					"addressV4",
					"addressV6",
					"clientAllowedIPs",
					"clientDns",
					"clientMtu",
					"persistentKeepalive",
					"enabled",
					"disabledReason",
					"expiresAt",
					"lastHandshakeAt",
					"lastEndpoint",
					"isOnline",
					"rxBytesTotal",
					"txBytesTotal",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICreateWgPeerBody": {
				"properties": {
					"interfaceId": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"userId": {
						"type": "string",
						"nullable": true,
						"description": "Держатель пира."
					},
					"publicKey": {
						"type": "string",
						"nullable": true,
						"description": "Импорт существующего клиента: его публичный ключ. Приватный ключ тогда\nне хранится, конфиг/QR не выпускаются."
					},
					"withPresharedKey": {
						"type": "boolean",
						"description": "Создавать ли preshared-ключ (по умолчанию да)."
					},
					"clientAllowedIPs": {
						"type": "string",
						"description": "AllowedIPs клиента (split tunnel); по умолчанию весь трафик."
					},
					"clientDns": {
						"type": "string",
						"nullable": true
					},
					"clientMtu": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"persistentKeepalive": {
						"type": "number",
						"format": "double"
					},
					"expiresAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"enabled": {
						"type": "boolean"
					}
				},
				"required": [
					"interfaceId",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IPaginatedDto_WgPeerDto_": {
				"description": "Страница по смещению: списки с известным общим числом.",
				"properties": {
					"items": {
						"items": {
							"$ref": "#/components/schemas/WgPeerDto"
						},
						"type": "array"
					},
					"total": {
						"type": "number",
						"format": "double",
						"description": "Всего элементов, подходящих под фильтр."
					},
					"offset": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"items",
					"total",
					"offset",
					"limit"
				],
				"type": "object",
				"additionalProperties": false
			},
			"WgPeerOptionDto": {
				"description": "Краткая запись для выпадающих списков.",
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"interfaceId": {
						"type": "string"
					},
					"addressV4": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"interfaceId",
					"addressV4"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUpdateWgPeerBody": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"clientAllowedIPs": {
						"type": "string"
					},
					"clientDns": {
						"type": "string",
						"nullable": true
					},
					"clientMtu": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"persistentKeepalive": {
						"type": "number",
						"format": "double"
					},
					"expiresAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"IAssignWgPeerBody": {
				"properties": {
					"userId": {
						"type": "string"
					}
				},
				"required": [
					"userId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IWgPeerQrDto": {
				"description": "QR-код клиентского конфига.",
				"properties": {
					"dataUrl": {
						"type": "string",
						"description": "PNG data-URL."
					}
				},
				"required": [
					"dataUrl"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IWgPeerLive": {
				"description": "Live-снимок пира.",
				"properties": {
					"peerId": {
						"type": "string"
					},
					"interfaceId": {
						"type": "string"
					},
					"nodeId": {
						"type": "string"
					},
					"userId": {
						"type": "string",
						"nullable": true
					},
					"online": {
						"type": "boolean"
					},
					"lastHandshakeAt": {
						"type": "string",
						"nullable": true
					},
					"endpoint": {
						"type": "string",
						"nullable": true
					},
					"rxTotal": {
						"type": "number",
						"format": "double"
					},
					"txTotal": {
						"type": "number",
						"format": "double"
					},
					"rxBps": {
						"type": "number",
						"format": "double"
					},
					"txBps": {
						"type": "number",
						"format": "double"
					},
					"ts": {
						"type": "string"
					}
				},
				"required": [
					"peerId",
					"interfaceId",
					"nodeId",
					"userId",
					"online",
					"lastHandshakeAt",
					"endpoint",
					"rxTotal",
					"txTotal",
					"rxBps",
					"txBps",
					"ts"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IWgSpeedPoint": {
				"description": "Точка короткой истории скорости (кольцевой ряд последних минут).",
				"properties": {
					"ts": {
						"type": "number",
						"format": "double",
						"description": "Момент сбора (unix ms)."
					},
					"rxBps": {
						"type": "number",
						"format": "double"
					},
					"txBps": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"ts",
					"rxBps",
					"txBps"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			},
			"apiKey": {
				"type": "apiKey",
				"in": "header",
				"name": "X-Api-Key",
				"description": "API-ключ: агент ноды (scope wg-agent) или интеграция"
			}
		}
	},
	"info": {
		"title": "wg-admin",
		"version": "1.0.1",
		"description": "WG Admin — админка WireGuard: бэкенд (Koa, tsoa, TypeORM, pg-boss) и агент нод на Go",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Epifanov Andrei",
			"email": "epifanovmd@gmail.com"
		}
	},
	"paths": {
...
		"/api/v1/wg/peers": {
			"post": {
				"operationId": "CreateWgPeer",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WgPeerDto"
								}
							}
						}
					},
					"default": {
						"description": "Ошибка",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponseDto"
								}
							}
						}
					}
				},
				"description": "Создать пира: ключи и IP выделяются автоматически; `publicKey` — импорт\nсуществующего клиента (его приватный ключ не хранится).",
				"summary": "Создание пира",
				"tags": [
					"WgPeer"
				],
				"security": [
					{
						"jwt": [
							"permission:wg:peer:create"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICreateWgPeerBody"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "ListWgPeers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPaginatedDto_WgPeerDto_"
								}
							}
						}
					},
					"default": {
						"description": "Ошибка",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponseDto"
								}
							}
						}
					}
				},
				"description": "Пиры с фильтрами. Без права `wg:peer:view` возвращаются только свои\nпиры (право `wg:peer:own`).",
				"summary": "Список пиров",
				"tags": [
					"WgPeer"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "interfaceId",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					},
					{
						"in": "query",
						"name": "nodeId",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					},
					{
						"in": "query",
						"name": "userId",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					},
					{
						"in": "query",
						"name": "enabled",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "online",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "query",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/v1/wg/peers/options": {
			"get": {
				"operationId": "WgPeerOptions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/WgPeerOptionDto"
									},
									"type": "array"
								}
							}
						}
					},
					"default": {
						"description": "Ошибка",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponseDto"
								}
							}
						}
					}
				},
				"description": "Краткий список пиров для выпадающих списков (в рамках прав).",
				"summary": "Пиры (options)",
				"tags": [
					"WgPeer"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/api/v1/wg/peers/{id}": {
			"get": {
				"operationId": "GetWgPeer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WgPeerDto"
								}
							}
						}
					},
					"default": {
						"description": "Ошибка",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponseDto"
								}
							}
						}
					}
				},
				"description": "Пир по id; чужой без права `wg:peer:view` — 404.",
				"summary": "Пир",
				"tags": [
					"WgPeer"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateWgPeer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WgPeerDto"
								}
							}
						}
					},
					"default": {
						"description": "Ошибка",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponseDto"
								}
							}
						}
					}
				},
				"description": "Изменить пира: переданные поля заменяются.",
				"summary": "Изменение пира",
				"tags": [
					"WgPeer"
				],
				"security": [
					{
						"jwt": [
							"permission:wg:peer:update"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUpdateWgPeerBody"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteWgPeer",
				"responses": {
					"204": {
						"description": "No Content"
					},
					"default": {
						"description": "Ошибка",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponseDto"
								}
							}
						}
					}
				},
				"description": "Удалить пира; агент снимет его с интерфейса.",
				"summary": "Удаление пира",
				"tags": [
					"WgPeer"
				],
				"security": [
					{
						"jwt": [
							"permission:wg:peer:delete"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				]
			}
		},
		"/api/v1/wg/peers/{id}/enable": {
			"post": {
				"operationId": "EnableWgPeer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WgPeerDto"
								}
							}
						}
					},
					"default": {
						"description": "Ошибка",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponseDto"
								}
							}
						}
					}
				},
				"description": "Включить пира. Держатель может включать свои пиры.",
				"summary": "Включение пира",
				"tags": [
					"WgPeer"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				]
			}
		},
		"/api/v1/wg/peers/{id}/disable": {
			"post": {
				"operationId": "DisableWgPeer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WgPeerDto"
								}
							}
						}
					},
					"default": {
						"description": "Ошибка",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponseDto"
								}
							}
						}
					}
				},
				"description": "Выключить пира. Держатель может выключать свои пиры.",
				"summary": "Выключение пира",
				"tags": [
					"WgPeer"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				]
			}
		},
		"/api/v1/wg/peers/{id}/psk/rotate": {
			"post": {
				"operationId": "RotateWgPeerPsk",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WgPeerDto"
								}
							}
						}
					},
					"default": {
						"description": "Ошибка",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponseDto"
								}
							}
						}
					}
				},
				"description": "Перевыпустить preshared-ключ; клиенту нужен новый конфиг.",
				"summary": "Ротация PSK",
				"tags": [
					"WgPeer"
				],
				"security": [
					{
						"jwt": [
							"permission:wg:peer:psk"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				]
			}
		},
		"/api/v1/wg/peers/{id}/psk": {
			"delete": {
				"operationId": "RemoveWgPeerPsk",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WgPeerDto"
								}
							}
						}
					},
					"default": {
						"description": "Ошибка",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponseDto"
								}
							}
						}
					}
				},
				"description": "Убрать preshared-ключ; клиенту нужен новый конфиг.",
				"summary": "Удаление PSK",
				"tags": [
					"WgPeer"
				],
				"security": [
					{
						"jwt": [
							"permission:wg:peer:psk"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				]
			}
		},
		"/api/v1/wg/peers/{id}/assign": {
			"post": {
				"operationId": "AssignWgPeer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WgPeerDto"
								}
							}
						}
					},
					"default": {
						"description": "Ошибка",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponseDto"
								}
							}
						}
					}
				},
				"description": "Назначить пира пользователю (он увидит его в «Моих пирах»).",
				"summary": "Назначение пира",
				"tags": [
					"WgPeer"
				],
				"security": [
					{
						"jwt": [
							"permission:wg:peer:assign"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IAssignWgPeerBody"
							}
						}
					}
				}
			}
		},
		"/api/v1/wg/peers/{id}/revoke": {
			"post": {
				"operationId": "RevokeWgPeer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WgPeerDto"
								}
							}
						}
					},
					"default": {
						"description": "Ошибка",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponseDto"
								}
							}
						}
					}
				},
				"description": "Отвязать пира от пользователя.",
				"summary": "Отвязка пира",
				"tags": [
					"WgPeer"
				],
				"security": [
					{
						"jwt": [
							"permission:wg:peer:assign"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				]
			}
		},
		"/api/v1/wg/peers/{id}/config": {
			"get": {
				"operationId": "WgPeerConfig",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"text/plain": {
								"schema": {
									"type": "string"
								}
							}
						}
					},
					"default": {
						"description": "Ошибка",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponseDto"
								}
							}
						}
					}
				},
				"description": "Клиентский конфиг `.conf` (attachment). Держатель получает конфиги\nсвоих пиров.",
				"summary": "Конфиг пира",
				"tags": [
					"WgPeer"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				]
			}
		},
		"/api/v1/wg/peers/{id}/qr": {
			"get": {
				"operationId": "WgPeerQr",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IWgPeerQrDto"
								}
							}
						}
					},
					"default": {
						"description": "Ошибка",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponseDto"
								}
							}
						}
					}
				},
				"description": "QR-код клиентского конфига (PNG data-URL).",
				"summary": "QR пира",
				"tags": [
					"WgPeer"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				]
			}
		},
...
		"/api/v1/wg/stats/current/peer/{peerId}": {
			"get": {
				"operationId": "WgStatsCurrentPeer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IWgPeerLive"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"default": {
						"description": "Ошибка",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponseDto"
								}
							}
						}
					}
				},
				"description": "Текущий live-снимок пира (для первой отрисовки, дальше — сокет).\nДержатель видит свои пиры; `null` — агент ещё не присылал статистику.",
				"summary": "Текущий снимок пира",
				"tags": [
					"WgStats"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "peerId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				]
			}
		},
...
		"/api/v1/wg/stats/window/peer/{peerId}": {
			"get": {
				"operationId": "WgStatsPeerWindow",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IWgSpeedPoint"
									},
									"type": "array"
								}
							}
						}
					},
					"default": {
						"description": "Ошибка",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponseDto"
								}
							}
						}
					}
				},
				"description": "Скорость пира за последние минуты (для первой отрисовки графика,\nдальше — сокет). Держатель видит свои пиры.",
				"summary": "Короткий ряд скорости пира",
				"tags": [
					"WgStats"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "peerId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:8181",
			"description": "Local dev"
		}
	]
}